Fix: Add AuthProvider to fix context-related crash
This commit introduces an `AuthProvider` and an `AuthContext` to manage the application's authentication state.

Previously, the authentication state was managed directly within the `DashboardLayout` component. This caused a crash in child components that tried to access the authentication context, as it was not being provided to them.

This commit refactors the authentication logic into a dedicated `AuthProvider` component. The `DashboardLayout` now wraps its children with this provider, ensuring that the authentication context is available to all components in the dashboard. This resolves the crash and improves the overall structure of the code by centralizing authentication management.
